+++ src/api/tasks/teacher_test.py
@@ -5,7 +5,7 @@ Created on 23/10/2025 at 14:29:26(+01:00).
 
 from codeforlife.tests import CeleryTestCase
 
-from .teacher import classes_per_teacher, teacher_logins
+from .teacher import classes_per_teacher
 
 # pylint: disable=missing-class-docstring
 
@@ -16,7 +16,3 @@ class TestTeacher(CeleryTestCase):
     def test_classes_per_teacher(self):
         """Assert the queryset returns the expected fields."""
         self.assert_data_warehouse_task(task=classes_per_teacher)
-
-    def test_teacher_logins(self):
-        """Assert the queryset returns the expected fields."""
-        self.assert_data_warehouse_task(task=teacher_logins)
